pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "jobboard"
authors = [{name = "Shako Rzayev", email = "rzayev.sehriyar@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE.md"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.9"
dependencies = [
    "fastapi == 0.90.1",
    "uvicorn >= 0.18.3",
    "dependency-injector >= 4.40.0",
    "psycopg2-binary >= 2.9.3",
    "python-dotenv >= 0.21.0",
    "passlib[bcrypt] >= 1.7.4",
    "python-jose >= 3.3.0",
    "sqlalchemy >= 1.4.41",
    "databases >= 0.6.1",
    "databases[sqlite] >= 0.6.1",
    "pydantic[email] >= 1.10.2",
    "jinja2 >=3.1.2",
    "python-multipart >=0.0.5",
    "alembic >=1.8.1",
    "loguru >=0.6.0",
    "httpx >=0.23.1",
    "opentelemetry-exporter-jaeger >=1.14.0",
    "opentelemetry-instrumentation-fastapi >=0.35b0",
    "opentelemetry-instrumentation-logging >=0.35b0",
    "opentelemetry-instrumentation-sqlalchemy >=0.35b0",
    "prometheus-fastapi-instrumentator >=5.9.1",
]

[project.optional-dependencies]
dev = [
    "black >=23.1.0",
    "isort >=5.12.0",
    "autoflake >=1.4",
    "flake8 >=6.0.0",
    "pre-commit >=2.21.0",
    "pytype >= 2022.12.15",
    "bandit >= 1.7.4",
    "flake8-bugbear >= 22.12.6",
    ]
test = [
    "pytest >=7.1.3",
    "pytest-cov >= 3.0.0",
]

[tool.isort]
profile = "black"
py_version = 39
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json"]
skip_glob = ["docs/*"]


[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.bandit]
skips = ["B311"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
testpaths = "tests"